Rename GrahpQL scalar type `bodyHTML` to `Html`
We want to be able to use Html like this in multiple places, not
restricted to bodies.

This also renames the body field on the comment to indicate the specific
rendered type of the body. This means we can use `body` to represent the
editable input/output format from Drupal later.